default imp floor assertion
